+++ src/irvine/oeis/a347/A347176.java
@@ -23,6 +23,6 @@ public class A347176 extends AbstractSequence {
   @Override
   public Z next() {
     ++mN;
-    return Integers.SINGLETON.sumdiv(mN, d -> Predicates.SQUARE.is((long) d) ? Z.valueOf((((((d & 1) + 1) & 1) == 0) ? 1 : -1) * IntegerUtils.sqrt(d)) : Z.ZERO);
+    return Integers.SINGLETON.sumdiv(mN, d -> Predicates.SQUARE.is(d) ? Z.valueOf((((((d & 1) + 1) & 1) == 0) ? 1 : -1) * IntegerUtils.sqrt(d)) : Z.ZERO);
   }
 }
